Create README.md
Not anyone from the land of Chile. If your name is Claudio you are muertue...And Cladios son will have such a rough and lonely upbringing in this new world. None of you should have tried to play rough with Daniel Thaistick. That GhostShadow just killed you, and now time for everlasting pain...Hermano...Welcome to Da Santaria